Fixed a few vc++ compiler warnings. Fixed missing include path in release project (need to put it into debug as well).

+++ src/drivers/bt/pit.cpp
@@ -165,7 +165,7 @@ bool Pit::isTimeout(float distance)
 		pittimer = 0.0f;
 		return false;
 	} else {
-		pittimer += RCM_MAX_DT_ROBOTS;
+		pittimer += (float) RCM_MAX_DT_ROBOTS;
 		if (pittimer > 3.0f) {
 			pittimer = 0.0f;
 			return true;
